+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -494,6 +494,28 @@ pub(crate) mod tests {
     }
 
     #[test]
+    fn send_max_single_step_proposed_transfer() {
+        testing::pool::send_max_single_step_proposed_transfer::<SaplingPoolTester>()
+    }
+
+    #[test]
+    #[cfg(feature = "transparent-inputs")]
+    fn fails_to_send_max_to_transparent_with_memo() {
+        testing::pool::fails_to_send_max_to_transparent_with_memo::<SaplingPoolTester>()
+    }
+
+    #[test]
+    fn send_max_proposal_fails_when_unconfirmed_funds_present() {
+        testing::pool::send_max_proposal_fails_when_unconfirmed_funds_present::<SaplingPoolTester>()
+    }
+
+    #[test]
+    #[cfg(feature = "transparent-inputs")]
+    fn send_max_multi_step_proposed_transfer() {
+        testing::pool::send_max_multi_step_proposed_transfer::<SaplingPoolTester>()
+    }
+
+    #[test]
     fn send_with_multiple_change_outputs() {
         testing::pool::send_with_multiple_change_outputs::<SaplingPoolTester>()
     }